+++ moneyed/classes.py
@@ -147,6 +147,16 @@ class Money(object):
             amount = (self.amount / force_decimal(other)).quantize(Decimal('1.00'), rounding=ROUND_HALF_UP)
             return self.__class__(amount=amount, currency=self.currency)
 
+    def round(self, ndigits=0):
+        """
+        Rounds the amount using the current ``Decimal`` rounding algorithm.
+        """
+        if ndigits is None:
+            ndigits = 0
+        return self.__class__(
+            amount=self.amount.quantize(Decimal('1e' + str(-ndigits))),
+            currency=self.currency)
+
     def __abs__(self):
         return self.__class__(amount=abs(self.amount), currency=self.currency)