+++ moneyed/test_moneyed_classes.py
@@ -136,7 +136,7 @@ class TestMoney:
         one_million_pln = Money('1000000', 'PLN')
         # Two decimal places by default
         assert format_money(one_million_pln, locale='pl_PL') == '1 000 000,00 zł'
-        assert format_money(self.one_million_bucks, locale='pl_PL') == '1 000 000,00 USD'
+        assert format_money(self.one_million_bucks, locale='pl_PL') == 'US$1 000 000,00'
         # No decimal point without fractional part
         assert format_money(one_million_pln, locale='pl_PL',
                             decimal_places=0) == '1 000 000 zł'
@@ -282,6 +282,34 @@ class TestMoney:
         assert (sum([Money(amount=1, currency=self.USD),
                      Money(amount=2, currency=self.USD)]) ==
                 Money(amount=3, currency=self.USD))
+
+    def test_round(self):
+        x = Money(amount='1234.33569', currency=self.USD)
+        assert x.round(-4) == Money(amount='0', currency=self.USD)
+        assert x.round(-3) == Money(amount='1000', currency=self.USD)
+        assert x.round(-2) == Money(amount='1200', currency=self.USD)
+        assert x.round(-1) == Money(amount='1230', currency=self.USD)
+        assert x.round(0) == Money(amount='1234', currency=self.USD)
+        assert x.round(None) == Money(amount='1234', currency=self.USD)
+        assert x.round(1) == Money(amount='1234.3', currency=self.USD)
+        assert x.round(2) == Money(amount='1234.34', currency=self.USD)
+        assert x.round(3) == Money(amount='1234.336', currency=self.USD)
+        assert x.round(4) == Money(amount='1234.3357', currency=self.USD)
+
+    def test_round_context_override(self):
+        import decimal
+
+        x = Money(amount='2.5', currency=self.USD)
+        assert x.round(0) == Money(amount=2, currency=self.USD)
+        x = Money(amount='3.5', currency=self.USD)
+        assert x.round(0) == Money(amount=4, currency=self.USD)
+
+        with decimal.localcontext() as ctx:
+            ctx.rounding = decimal.ROUND_HALF_UP
+            x = Money(amount='2.5', currency=self.USD)
+            assert x.round(0) == Money(amount=3, currency=self.USD)
+            x = Money(amount='3.5', currency=self.USD)
+            assert x.round(0) == Money(amount=4, currency=self.USD)
 
     def test_arithmetic_operations_return_real_subclass_instance(self):
         """